@@ -3,7 +3,9 @@
   "version": "1.0.0",
   "description": "Website for sharing meals",
   "engines": {
-    "node": "8.11.1"
+
+    "node": "^21.6.2"
+
   },
   "scripts": {
     "build": "webpack --mode production",
@@ -15,6 +17,9 @@
   "author": "Benjamin Hughes",
   "license": "MIT",
   "proxy": "http://localhost:5000",
+
+  "type": "module",
+
   "dependencies": {
     "@babel/core": "^7.12.3",
     "@babel/plugin-proposal-class-properties": "^7.12.1",
@@ -30,13 +35,17 @@
     "cors": "^2.8.4",
     "css-loader": "^5.0.0",
     "debug": "~2.6.0",
-    "dotenv": "^4.0.0",
+
+    "dotenv": "^16.4.5",
+
     "dotenv-webpack": "^7.1.0",
     "express": "^4.16.4",
     "express-brute": "^1.0.1",
     "express-form-data": "^2.0.1",
     "html-webpack-plugin": "^4.3.0",
-    "knex": "^0.21.21",
+
+    "knex": "^0.21.1",
+
     "multer": "^1.3.0",
     "mysql2": "^2.1.0",
     "path": "^0.12.7",
