@@ -1,13 +1,23 @@
-const express = require("express");
+import express from "express";
+import path from "path";
+import mealsRouter from "./api/meals.js";
+import reservationsRouter from "./api/reservations.js"
+import cors from "cors";
+import { fileURLToPath } from 'url';
+import { dirname } from 'path';
+import knex from "knex";
+import { error } from "console";
+
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
+
 const app = express();
 const router = express.Router();
-const path = require("path");
 
-const mealsRouter = require("./api/meals");
 const buildPath = path.join(__dirname, "../../dist");
 const port = process.env.PORT || 3000;
-const cors = require("cors");
-const knex = require("./database");
+
 
 // For week4 no need to look into this!
 // Serve the built client html
@@ -20,22 +30,34 @@ app.use(express.json());
 
 app.use(cors());
 
-router.use("/meals", mealsRouter);
+app.use("/api/meals", mealsRouter);
+app.use("/api/reservations", reservationsRouter)
+
+if (process.env.API_PATH) {
+  app.use(process.env.API_PATH, router);
+} else {
+  throw "API_PATH is not set. Remember to set it in your .env file"
+}
+
+// for the frontend. Will first be covered in the react class
+app.use("*", (req, res) => {
+  res.sendFile(path.join(`${buildPath}/index.html`));
+});
 
 // Respond with all meals in the future (relative to the when datetime)
-app.get("/future-meals", async(req, res) => {
+app.get("/future-meals", async (req, res) => {
   try {
     const futureMeals = await knex("Meal")
       .select()
-      .where('when', '>', new Date());
+      .where("when", ">", new Date());
     if (futureMeals.length === 0) {
       res.status(404).send("The future meals is empty");
       return;
     }
     res.status(200).json(futureMeals);
-  } catch(error) {
-      console.error(error);
-      res.status(500).send("Something went wrong");
+  } catch (error) {
+    console.log(error);
+    res.status(500).send("Something went wrong");
   }
 });
 
@@ -43,9 +65,7 @@ app.get("/future-meals", async(req, res) => {
 app.get("/past-meals", async (req, res) => {
   try {
     const date = new Date();
-    const pastMeals = await knex("Meal")
-      .select()
-      .where("when", "<", date);
+    const pastMeals = await knex("Meal").select().where("when", "<", date);
     res.status(200).json(pastMeals);
   } catch (error) {
     console.error(error);
@@ -56,9 +76,7 @@ app.get("/past-meals", async (req, res) => {
 //Respond with all meals sorted by ID
 app.get("/all-meals", async (req, res) => {
   try {
-    const allMeals = await knex("Meal")
-      .select()
-      .orderBy('id');
+    const allMeals = await knex("Meal").select().orderBy("id");
     res.status(200).json(allMeals);
   } catch (error) {
     console.error(error);
@@ -83,8 +101,8 @@ app.get("/first-meal", async (req, res) => {
 app.get("/last-meal", async (req, res) => {
   try {
 
-    const firstMeal = await knex("Meal").select().orderBy("id", 'desc').first();
-  res.status(200).json(lastMeal);
+    const lastMeal = await knex("Meal").select().orderBy("id", "desc").first();
+    res.status(200).json(lastMeal);
 
   } catch (error) {
     console.error(error);
@@ -92,16 +110,4 @@ app.get("/last-meal", async (req, res) => {
   }
 });
 
-
-if (process.env.API_PATH) {
-  app.use(process.env.API_PATH, router);
-} else {
-  throw "API_PATH is not set. Remember to set it in your .env file"
-}
-
-// for the frontend. Will first be covered in the react class
-app.use("*", (req, res) => {
-  res.sendFile(path.join(`${buildPath}/index.html`));
-});
-
-module.exports = app;
+export default app;
