@@ -1,4 +1,6 @@
-const app = require("./app");
+
+import app from "./app.js";
+
 
 const port = parseInt(process.env.PORT, 10) || process.env.API_PORT;
 
