@@ -1,6 +1,8 @@
 import React from "react";
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
-import TestComponent from "./components/TestComponent/TestComponent";
+
+import TestComponent from "./components/TestComponent/TestComponent.js";
+
 
 function App() {
   return (
